requirements.txt
fastapi
uvicorn
langchain
langgraph
langchain-community
langchain-ollama
langchain-openai
langchain-google-genai
pydantic
pypdf
psycopg2-binary
pgvector
sqlalchemy
sqlmodel
jupyter
notebook
pandas
numpy
scikit-learn
python-multipart
python-dotenv
unstructured
pdf2image
pdfminer.six
pymupdf
rank_bm25
sentence-transformers
pytesseract
pdf2image
Pillow
redis
celery
passlib[argon2]
argon2-cffi
python-jose[cryptography]
